Create Random File Paths Using English Words

Create Random File Paths Using English Words

random_paths(max_depth = 5, min_chars = 5, max_elements = 10,
  depth_to_leaf_weight = function(depth) 1.2^depth)

Arguments

max_depth
maximum path depth

min_chars
least number of characters per folder or file name

max_elements
maximum number of elements (files or subfolders) in a folder

depth_to_leaf_weight
function that calculates a weight from the given path depth. The weight is used to increase the probability of a folder element to be a file and not a subdirectory. By default the weight is calculated by 1.2^depth, i.e. for a folder in depth 10 it is about six times (1.2^10 = 6.19) more probable of its elements to be files instead of subfolders

Examples

paths <- kwb.pathdict::random_paths(max_depth = 5)

#>  Creating node  1 / 9 
#>  Creating node  2 / 9 
#>    Creating node  1 / 5 
#>    Creating node  2 / 5 
#>    Creating node  3 / 5 
#>    Creating node  4 / 5 
#>    Creating node  5 / 5 
#>      Creating node  1 / 7 
#>      Creating node  2 / 7 
#>      Creating node  3 / 7 
#>      Creating node  4 / 7 
#>        Creating node  1 / 1 
#>          Creating node  1 / 10 
#>          Creating node  2 / 10 
#>          Creating node  3 / 10 
#>          Creating node  4 / 10 
#>          Creating node  5 / 10 
#>          Creating node  6 / 10 
#>          Creating node  7 / 10 
#>          Creating node  8 / 10 
#>          Creating node  9 / 10 
#>          Creating node  10 / 10 
#>      Creating node  5 / 7 
#>      Creating node  6 / 7 
#>      Creating node  7 / 7 
#>        Creating node  1 / 3 
#>        Creating node  2 / 3 
#>          Creating node  1 / 10 
#>          Creating node  2 / 10 
#>          Creating node  3 / 10 
#>          Creating node  4 / 10 
#>          Creating node  5 / 10 
#>          Creating node  6 / 10 
#>          Creating node  7 / 10 
#>          Creating node  8 / 10 
#>          Creating node  9 / 10 
#>          Creating node  10 / 10 
#>        Creating node  3 / 3 
#>  Creating node  3 / 9 
#>    Creating node  1 / 9 
#>    Creating node  2 / 9 
#>    Creating node  3 / 9 
#>      Creating node  1 / 10 
#>      Creating node  2 / 10 
#>        Creating node  1 / 3 
#>        Creating node  2 / 3 
#>        Creating node  3 / 3 
#>      Creating node  3 / 10 
#>        Creating node  1 / 8 
#>          Creating node  1 / 5 
#>          Creating node  2 / 5 
#>          Creating node  3 / 5 
#>          Creating node  4 / 5 
#>          Creating node  5 / 5 
#>        Creating node  2 / 8 
#>          Creating node  1 / 4 
#>          Creating node  2 / 4 
#>          Creating node  3 / 4 
#>          Creating node  4 / 4 
#>        Creating node  3 / 8 
#>        Creating node  4 / 8 
#>        Creating node  5 / 8 
#>          Creating node  1 / 10 
#>          Creating node  2 / 10 
#>          Creating node  3 / 10 
#>          Creating node  4 / 10 
#>          Creating node  5 / 10 
#>          Creating node  6 / 10 
#>          Creating node  7 / 10 
#>          Creating node  8 / 10 
#>          Creating node  9 / 10 
#>          Creating node  10 / 10 
#>        Creating node  6 / 8 
#>          Creating node  1 / 1 
#>        Creating node  7 / 8 
#>        Creating node  8 / 8 
#>      Creating node  4 / 10 
#>      Creating node  5 / 10 
#>        Creating node  1 / 8 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  2 / 8 
#>        Creating node  3 / 8 
#>          Creating node  1 / 1 
#>        Creating node  4 / 8 
#>        Creating node  5 / 8 
#>        Creating node  6 / 8 
#>        Creating node  7 / 8 
#>        Creating node  8 / 8 
#>      Creating node  6 / 10 
#>      Creating node  7 / 10 
#>        Creating node  1 / 8 
#>          Creating node  1 / 1 
#>        Creating node  2 / 8 
#>        Creating node  3 / 8 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  4 / 8 
#>        Creating node  5 / 8 
#>        Creating node  6 / 8 
#>          Creating node  1 / 2 
#>          Creating node  2 / 2 
#>        Creating node  7 / 8 
#>        Creating node  8 / 8 
#>          Creating node  1 / 10 
#>          Creating node  2 / 10 
#>          Creating node  3 / 10 
#>          Creating node  4 / 10 
#>          Creating node  5 / 10 
#>          Creating node  6 / 10 
#>          Creating node  7 / 10 
#>          Creating node  8 / 10 
#>          Creating node  9 / 10 
#>          Creating node  10 / 10 
#>      Creating node  8 / 10 
#>      Creating node  9 / 10 
#>        Creating node  1 / 5 
#>        Creating node  2 / 5 
#>        Creating node  3 / 5 
#>        Creating node  4 / 5 
#>        Creating node  5 / 5 
#>      Creating node  10 / 10 
#>        Creating node  1 / 2 
#>        Creating node  2 / 2 
#>    Creating node  4 / 9 
#>    Creating node  5 / 9 
#>    Creating node  6 / 9 
#>    Creating node  7 / 9 
#>      Creating node  1 / 8 
#>      Creating node  2 / 8 
#>        Creating node  1 / 7 
#>        Creating node  2 / 7 
#>        Creating node  3 / 7 
#>          Creating node  1 / 6 
#>          Creating node  2 / 6 
#>          Creating node  3 / 6 
#>          Creating node  4 / 6 
#>          Creating node  5 / 6 
#>          Creating node  6 / 6 
#>        Creating node  4 / 7 
#>        Creating node  5 / 7 
#>        Creating node  6 / 7 
#>        Creating node  7 / 7 
#>      Creating node  3 / 8 
#>        Creating node  1 / 1 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>      Creating node  4 / 8 
#>        Creating node  1 / 6 
#>          Creating node  1 / 2 
#>          Creating node  2 / 2 
#>        Creating node  2 / 6 
#>        Creating node  3 / 6 
#>        Creating node  4 / 6 
#>        Creating node  5 / 6 
#>        Creating node  6 / 6 
#>      Creating node  5 / 8 
#>      Creating node  6 / 8 
#>      Creating node  7 / 8 
#>      Creating node  8 / 8 
#>    Creating node  8 / 9 
#>      Creating node  1 / 8 
#>      Creating node  2 / 8 
#>        Creating node  1 / 7 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  2 / 7 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  3 / 7 
#>          Creating node  1 / 8 
#>          Creating node  2 / 8 
#>          Creating node  3 / 8 
#>          Creating node  4 / 8 
#>          Creating node  5 / 8 
#>          Creating node  6 / 8 
#>          Creating node  7 / 8 
#>          Creating node  8 / 8 
#>        Creating node  4 / 7 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  5 / 7 
#>        Creating node  6 / 7 
#>        Creating node  7 / 7 
#>      Creating node  3 / 8 
#>        Creating node  1 / 4 
#>        Creating node  2 / 4 
#>        Creating node  3 / 4 
#>          Creating node  1 / 8 
#>          Creating node  2 / 8 
#>          Creating node  3 / 8 
#>          Creating node  4 / 8 
#>          Creating node  5 / 8 
#>          Creating node  6 / 8 
#>          Creating node  7 / 8 
#>          Creating node  8 / 8 
#>        Creating node  4 / 4 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>      Creating node  4 / 8 
#>      Creating node  5 / 8 
#>      Creating node  6 / 8 
#>        Creating node  1 / 7 
#>        Creating node  2 / 7 
#>        Creating node  3 / 7 
#>        Creating node  4 / 7 
#>        Creating node  5 / 7 
#>          Creating node  1 / 4 
#>          Creating node  2 / 4 
#>          Creating node  3 / 4 
#>          Creating node  4 / 4 
#>        Creating node  6 / 7 
#>        Creating node  7 / 7 
#>      Creating node  7 / 8 
#>      Creating node  8 / 8 
#>    Creating node  9 / 9 
#>      Creating node  1 / 2 
#>        Creating node  1 / 9 
#>        Creating node  2 / 9 
#>        Creating node  3 / 9 
#>        Creating node  4 / 9 
#>        Creating node  5 / 9 
#>        Creating node  6 / 9 
#>          Creating node  1 / 10 
#>          Creating node  2 / 10 
#>          Creating node  3 / 10 
#>          Creating node  4 / 10 
#>          Creating node  5 / 10 
#>          Creating node  6 / 10 
#>          Creating node  7 / 10 
#>          Creating node  8 / 10 
#>          Creating node  9 / 10 
#>          Creating node  10 / 10 
#>        Creating node  7 / 9 
#>          Creating node  1 / 10 
#>          Creating node  2 / 10 
#>          Creating node  3 / 10 
#>          Creating node  4 / 10 
#>          Creating node  5 / 10 
#>          Creating node  6 / 10 
#>          Creating node  7 / 10 
#>          Creating node  8 / 10 
#>          Creating node  9 / 10 
#>          Creating node  10 / 10 
#>        Creating node  8 / 9 
#>          Creating node  1 / 1 
#>        Creating node  9 / 9 
#>      Creating node  2 / 2 
#>  Creating node  4 / 9 
#>    Creating node  1 / 10 
#>    Creating node  2 / 10 
#>      Creating node  1 / 4 
#>        Creating node  1 / 9 
#>          Creating node  1 / 10 
#>          Creating node  2 / 10 
#>          Creating node  3 / 10 
#>          Creating node  4 / 10 
#>          Creating node  5 / 10 
#>          Creating node  6 / 10 
#>          Creating node  7 / 10 
#>          Creating node  8 / 10 
#>          Creating node  9 / 10 
#>          Creating node  10 / 10 
#>        Creating node  2 / 9 
#>        Creating node  3 / 9 
#>        Creating node  4 / 9 
#>          Creating node  1 / 4 
#>          Creating node  2 / 4 
#>          Creating node  3 / 4 
#>          Creating node  4 / 4 
#>        Creating node  5 / 9 
#>          Creating node  1 / 7 
#>          Creating node  2 / 7 
#>          Creating node  3 / 7 
#>          Creating node  4 / 7 
#>          Creating node  5 / 7 
#>          Creating node  6 / 7 
#>          Creating node  7 / 7 
#>        Creating node  6 / 9 
#>        Creating node  7 / 9 
#>        Creating node  8 / 9 
#>        Creating node  9 / 9 
#>      Creating node  2 / 4 
#>        Creating node  1 / 8 
#>        Creating node  2 / 8 
#>          Creating node  1 / 5 
#>          Creating node  2 / 5 
#>          Creating node  3 / 5 
#>          Creating node  4 / 5 
#>          Creating node  5 / 5 
#>        Creating node  3 / 8 
#>        Creating node  4 / 8 
#>          Creating node  1 / 2 
#>          Creating node  2 / 2 
#>        Creating node  5 / 8 
#>        Creating node  6 / 8 
#>          Creating node  1 / 8 
#>          Creating node  2 / 8 
#>          Creating node  3 / 8 
#>          Creating node  4 / 8 
#>          Creating node  5 / 8 
#>          Creating node  6 / 8 
#>          Creating node  7 / 8 
#>          Creating node  8 / 8 
#>        Creating node  7 / 8 
#>          Creating node  1 / 4 
#>          Creating node  2 / 4 
#>          Creating node  3 / 4 
#>          Creating node  4 / 4 
#>        Creating node  8 / 8 
#>      Creating node  3 / 4 
#>        Creating node  1 / 6 
#>          Creating node  1 / 6 
#>          Creating node  2 / 6 
#>          Creating node  3 / 6 
#>          Creating node  4 / 6 
#>          Creating node  5 / 6 
#>          Creating node  6 / 6 
#>        Creating node  2 / 6 
#>          Creating node  1 / 6 
#>          Creating node  2 / 6 
#>          Creating node  3 / 6 
#>          Creating node  4 / 6 
#>          Creating node  5 / 6 
#>          Creating node  6 / 6 
#>        Creating node  3 / 6 
#>        Creating node  4 / 6 
#>        Creating node  5 / 6 
#>        Creating node  6 / 6 
#>      Creating node  4 / 4 
#>    Creating node  3 / 10 
#>      Creating node  1 / 2 
#>      Creating node  2 / 2 
#>        Creating node  1 / 9 
#>        Creating node  2 / 9 
#>        Creating node  3 / 9 
#>        Creating node  4 / 9 
#>          Creating node  1 / 5 
#>          Creating node  2 / 5 
#>          Creating node  3 / 5 
#>          Creating node  4 / 5 
#>          Creating node  5 / 5 
#>        Creating node  5 / 9 
#>        Creating node  6 / 9 
#>        Creating node  7 / 9 
#>        Creating node  8 / 9 
#>        Creating node  9 / 9 
#>          Creating node  1 / 7 
#>          Creating node  2 / 7 
#>          Creating node  3 / 7 
#>          Creating node  4 / 7 
#>          Creating node  5 / 7 
#>          Creating node  6 / 7 
#>          Creating node  7 / 7 
#>    Creating node  4 / 10 
#>    Creating node  5 / 10 
#>    Creating node  6 / 10 
#>    Creating node  7 / 10 
#>      Creating node  1 / 5 
#>      Creating node  2 / 5 
#>        Creating node  1 / 6 
#>          Creating node  1 / 2 
#>          Creating node  2 / 2 
#>        Creating node  2 / 6 
#>          Creating node  1 / 4 
#>          Creating node  2 / 4 
#>          Creating node  3 / 4 
#>          Creating node  4 / 4 
#>        Creating node  3 / 6 
#>        Creating node  4 / 6 
#>          Creating node  1 / 6 
#>          Creating node  2 / 6 
#>          Creating node  3 / 6 
#>          Creating node  4 / 6 
#>          Creating node  5 / 6 
#>          Creating node  6 / 6 
#>        Creating node  5 / 6 
#>          Creating node  1 / 3 
#>          Creating node  2 / 3 
#>          Creating node  3 / 3 
#>        Creating node  6 / 6 
#>          Creating node  1 / 5 
#>          Creating node  2 / 5 
#>          Creating node  3 / 5 
#>          Creating node  4 / 5 
#>          Creating node  5 / 5 
#>      Creating node  3 / 5 
#>      Creating node  4 / 5 
#>      Creating node  5 / 5 
#>        Creating node  1 / 6 
#>          Creating node  1 / 2 
#>          Creating node  2 / 2 
#>        Creating node  2 / 6 
#>        Creating node  3 / 6 
#>          Creating node  1 / 3 
#>          Creating node  2 / 3 
#>          Creating node  3 / 3 
#>        Creating node  4 / 6 
#>        Creating node  5 / 6 
#>        Creating node  6 / 6 
#>    Creating node  8 / 10 
#>    Creating node  9 / 10 
#>      Creating node  1 / 1 
#>    Creating node  10 / 10 
#>  Creating node  5 / 9 
#>    Creating node  1 / 7 
#>      Creating node  1 / 6 
#>      Creating node  2 / 6 
#>      Creating node  3 / 6 
#>      Creating node  4 / 6 
#>      Creating node  5 / 6 
#>      Creating node  6 / 6 
#>    Creating node  2 / 7 
#>    Creating node  3 / 7 
#>      Creating node  1 / 6 
#>      Creating node  2 / 6 
#>        Creating node  1 / 4 
#>          Creating node  1 / 6 
#>          Creating node  2 / 6 
#>          Creating node  3 / 6 
#>          Creating node  4 / 6 
#>          Creating node  5 / 6 
#>          Creating node  6 / 6 
#>        Creating node  2 / 4 
#>        Creating node  3 / 4 
#>        Creating node  4 / 4 
#>      Creating node  3 / 6 
#>        Creating node  1 / 9 
#>          Creating node  1 / 2 
#>          Creating node  2 / 2 
#>        Creating node  2 / 9 
#>        Creating node  3 / 9 
#>          Creating node  1 / 1 
#>        Creating node  4 / 9 
#>          Creating node  1 / 7 
#>          Creating node  2 / 7 
#>          Creating node  3 / 7 
#>          Creating node  4 / 7 
#>          Creating node  5 / 7 
#>          Creating node  6 / 7 
#>          Creating node  7 / 7 
#>        Creating node  5 / 9 
#>          Creating node  1 / 10 
#>          Creating node  2 / 10 
#>          Creating node  3 / 10 
#>          Creating node  4 / 10 
#>          Creating node  5 / 10 
#>          Creating node  6 / 10 
#>          Creating node  7 / 10 
#>          Creating node  8 / 10 
#>          Creating node  9 / 10 
#>          Creating node  10 / 10 
#>        Creating node  6 / 9 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  7 / 9 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  8 / 9 
#>        Creating node  9 / 9 
#>          Creating node  1 / 1 
#>      Creating node  4 / 6 
#>      Creating node  5 / 6 
#>        Creating node  1 / 10 
#>          Creating node  1 / 10 
#>          Creating node  2 / 10 
#>          Creating node  3 / 10 
#>          Creating node  4 / 10 
#>          Creating node  5 / 10 
#>          Creating node  6 / 10 
#>          Creating node  7 / 10 
#>          Creating node  8 / 10 
#>          Creating node  9 / 10 
#>          Creating node  10 / 10 
#>        Creating node  2 / 10 
#>          Creating node  1 / 1 
#>        Creating node  3 / 10 
#>          Creating node  1 / 7 
#>          Creating node  2 / 7 
#>          Creating node  3 / 7 
#>          Creating node  4 / 7 
#>          Creating node  5 / 7 
#>          Creating node  6 / 7 
#>          Creating node  7 / 7 
#>        Creating node  4 / 10 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  5 / 10 
#>        Creating node  6 / 10 
#>        Creating node  7 / 10 
#>          Creating node  1 / 8 
#>          Creating node  2 / 8 
#>          Creating node  3 / 8 
#>          Creating node  4 / 8 
#>          Creating node  5 / 8 
#>          Creating node  6 / 8 
#>          Creating node  7 / 8 
#>          Creating node  8 / 8 
#>        Creating node  8 / 10 
#>          Creating node  1 / 6 
#>          Creating node  2 / 6 
#>          Creating node  3 / 6 
#>          Creating node  4 / 6 
#>          Creating node  5 / 6 
#>          Creating node  6 / 6 
#>        Creating node  9 / 10 
#>        Creating node  10 / 10 
#>      Creating node  6 / 6 
#>        Creating node  1 / 10 
#>        Creating node  2 / 10 
#>          Creating node  1 / 1 
#>        Creating node  3 / 10 
#>        Creating node  4 / 10 
#>        Creating node  5 / 10 
#>        Creating node  6 / 10 
#>        Creating node  7 / 10 
#>        Creating node  8 / 10 
#>        Creating node  9 / 10 
#>        Creating node  10 / 10 
#>    Creating node  4 / 7 
#>      Creating node  1 / 2 
#>      Creating node  2 / 2 
#>    Creating node  5 / 7 
#>      Creating node  1 / 9 
#>      Creating node  2 / 9 
#>      Creating node  3 / 9 
#>      Creating node  4 / 9 
#>      Creating node  5 / 9 
#>      Creating node  6 / 9 
#>      Creating node  7 / 9 
#>      Creating node  8 / 9 
#>      Creating node  9 / 9 
#>        Creating node  1 / 10 
#>          Creating node  1 / 6 
#>          Creating node  2 / 6 
#>          Creating node  3 / 6 
#>          Creating node  4 / 6 
#>          Creating node  5 / 6 
#>          Creating node  6 / 6 
#>        Creating node  2 / 10 
#>        Creating node  3 / 10 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  4 / 10 
#>        Creating node  5 / 10 
#>          Creating node  1 / 8 
#>          Creating node  2 / 8 
#>          Creating node  3 / 8 
#>          Creating node  4 / 8 
#>          Creating node  5 / 8 
#>          Creating node  6 / 8 
#>          Creating node  7 / 8 
#>          Creating node  8 / 8 
#>        Creating node  6 / 10 
#>        Creating node  7 / 10 
#>        Creating node  8 / 10 
#>        Creating node  9 / 10 
#>        Creating node  10 / 10 
#>    Creating node  6 / 7 
#>    Creating node  7 / 7 
#>  Creating node  6 / 9 
#>    Creating node  1 / 7 
#>      Creating node  1 / 7 
#>      Creating node  2 / 7 
#>      Creating node  3 / 7 
#>      Creating node  4 / 7 
#>      Creating node  5 / 7 
#>      Creating node  6 / 7 
#>        Creating node  1 / 7 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  2 / 7 
#>          Creating node  1 / 2 
#>          Creating node  2 / 2 
#>        Creating node  3 / 7 
#>        Creating node  4 / 7 
#>        Creating node  5 / 7 
#>        Creating node  6 / 7 
#>        Creating node  7 / 7 
#>      Creating node  7 / 7 
#>    Creating node  2 / 7 
#>    Creating node  3 / 7 
#>    Creating node  4 / 7 
#>    Creating node  5 / 7 
#>    Creating node  6 / 7 
#>      Creating node  1 / 6 
#>      Creating node  2 / 6 
#>      Creating node  3 / 6 
#>      Creating node  4 / 6 
#>        Creating node  1 / 9 
#>        Creating node  2 / 9 
#>          Creating node  1 / 2 
#>          Creating node  2 / 2 
#>        Creating node  3 / 9 
#>        Creating node  4 / 9 
#>        Creating node  5 / 9 
#>        Creating node  6 / 9 
#>        Creating node  7 / 9 
#>          Creating node  1 / 4 
#>          Creating node  2 / 4 
#>          Creating node  3 / 4 
#>          Creating node  4 / 4 
#>        Creating node  8 / 9 
#>          Creating node  1 / 6 
#>          Creating node  2 / 6 
#>          Creating node  3 / 6 
#>          Creating node  4 / 6 
#>          Creating node  5 / 6 
#>          Creating node  6 / 6 
#>        Creating node  9 / 9 
#>      Creating node  5 / 6 
#>      Creating node  6 / 6 
#>    Creating node  7 / 7 
#>      Creating node  1 / 7 
#>      Creating node  2 / 7 
#>      Creating node  3 / 7 
#>      Creating node  4 / 7 
#>      Creating node  5 / 7 
#>        Creating node  1 / 8 
#>          Creating node  1 / 4 
#>          Creating node  2 / 4 
#>          Creating node  3 / 4 
#>          Creating node  4 / 4 
#>        Creating node  2 / 8 
#>        Creating node  3 / 8 
#>        Creating node  4 / 8 
#>        Creating node  5 / 8 
#>        Creating node  6 / 8 
#>        Creating node  7 / 8 
#>        Creating node  8 / 8 
#>          Creating node  1 / 3 
#>          Creating node  2 / 3 
#>          Creating node  3 / 3 
#>      Creating node  6 / 7 
#>        Creating node  1 / 10 
#>        Creating node  2 / 10 
#>        Creating node  3 / 10 
#>          Creating node  1 / 8 
#>          Creating node  2 / 8 
#>          Creating node  3 / 8 
#>          Creating node  4 / 8 
#>          Creating node  5 / 8 
#>          Creating node  6 / 8 
#>          Creating node  7 / 8 
#>          Creating node  8 / 8 
#>        Creating node  4 / 10 
#>        Creating node  5 / 10 
#>        Creating node  6 / 10 
#>        Creating node  7 / 10 
#>        Creating node  8 / 10 
#>        Creating node  9 / 10 
#>        Creating node  10 / 10 
#>          Creating node  1 / 1 
#>      Creating node  7 / 7 
#>        Creating node  1 / 7 
#>        Creating node  2 / 7 
#>          Creating node  1 / 7 
#>          Creating node  2 / 7 
#>          Creating node  3 / 7 
#>          Creating node  4 / 7 
#>          Creating node  5 / 7 
#>          Creating node  6 / 7 
#>          Creating node  7 / 7 
#>        Creating node  3 / 7 
#>        Creating node  4 / 7 
#>        Creating node  5 / 7 
#>          Creating node  1 / 2 
#>          Creating node  2 / 2 
#>        Creating node  6 / 7 
#>          Creating node  1 / 9 
#>          Creating node  2 / 9 
#>          Creating node  3 / 9 
#>          Creating node  4 / 9 
#>          Creating node  5 / 9 
#>          Creating node  6 / 9 
#>          Creating node  7 / 9 
#>          Creating node  8 / 9 
#>          Creating node  9 / 9 
#>        Creating node  7 / 7 
#>  Creating node  7 / 9 
#>  Creating node  8 / 9 
#>  Creating node  9 / 9 

table(lengths(kwb.file::split_paths(paths)))

#> Splitting paths ... ok. (0.00s) 

#> 
#>   2   3   4   5   6 
#>   4  22  59 134 422