Create Random File Paths Using English Words

Create Random File Paths Using English Words

random_paths(max_depth = 5, min_chars = 5, max_elements = 10,
  depth_to_leaf_weight = function(depth) 1.2^depth)

Arguments

max_depth
maximum path depth

min_chars
least number of characters per folder or file name

max_elements
maximum number of elements (files or subfolders) in a folder

depth_to_leaf_weight
function that calculates a weight from the given path depth. The weight is used to increase the probability of a folder element to be a file and not a subdirectory. By default the weight is calculated by 1.2^depth, i.e. for a folder in depth 10 it is about six times (1.2^10 = 6.19) more probable of its elements to be files instead of subfolders

Examples

# Make this example reproducible
set.seed(12059)

# Create random paths
paths <- kwb.pathdict::random_paths(max_depth = 5)

# Show the random paths
paths

#>   [1] "dance/chair/shine/afraid.xls"                     
#>   [2] "dance/chair/shine/other/determine.R"              
#>   [3] "dance/chair/shine/other/since.pdf"                
#>   [4] "dance/chair/shine/other/garden.xls"               
#>   [5] "dance/chair/shine/other/shout/while.png"          
#>   [6] "dance/chair/shine/other/segment/contain.R"        
#>   [7] "dance/chair/shine/other/segment/middle.png"       
#>   [8] "dance/chair/shine/other/segment/thought.jpg"      
#>   [9] "dance/chair/shine/other/segment/shine.pdf"        
#>  [10] "dance/chair/shine/other/segment/father.xls"       
#>  [11] "dance/chair/shine/other/segment/village.xls"      
#>  [12] "dance/chair/shine/other/segment/character.doc"    
#>  [13] "dance/chair/shine/other/segment/danger.doc"       
#>  [14] "dance/chair/shine/other/segment/electric.jpg"     
#>  [15] "dance/chair/shine/other/death.R"                  
#>  [16] "dance/chair/shine/other/river/ground.xls"         
#>  [17] "dance/chair/shine/other/river/cover.png"          
#>  [18] "dance/chair/shine/other/river/problem.R"          
#>  [19] "dance/chair/shine/other/river/twenty.png"         
#>  [20] "dance/chair/shine/other/river/voice.xls"          
#>  [21] "dance/chair/shine/other/river/corner.png"         
#>  [22] "dance/chair/shine/other/river/state.xls"          
#>  [23] "dance/chair/shine/other/morning/visit.jpg"        
#>  [24] "dance/chair/shine/other/morning/observe.png"      
#>  [25] "dance/chair/shine/other/morning/century.R"        
#>  [26] "dance/chair/shine/other/morning/every.pdf"        
#>  [27] "dance/chair/shine/other/morning/company.doc"      
#>  [28] "dance/chair/shine/other/morning/still.xls"        
#>  [29] "dance/chair/shine/other/effect/spell.jpg"         
#>  [30] "dance/chair/shine/other/effect/month.png"         
#>  [31] "dance/chair/shine/other/effect/support.png"       
#>  [32] "dance/chair/shine/other/effect/class.png"         
#>  [33] "dance/chair/shine/other/effect/station.R"         
#>  [34] "dance/chair/shine/other/effect/sleep.xls"         
#>  [35] "dance/chair/shine/teeth/family/similar.R"         
#>  [36] "dance/chair/shine/teeth/family/probable.jpg"      
#>  [37] "dance/chair/shine/teeth/family/paragraph.png"     
#>  [38] "dance/chair/shine/teeth/family/result.png"        
#>  [39] "dance/chair/shine/black.pdf"                      
#>  [40] "dance/chair/shine/house.png"                      
#>  [41] "dance/chair/shine/particular/certain/correct.xls" 
#>  [42] "dance/chair/shine/particular/certain/catch.jpg"   
#>  [43] "dance/chair/shine/particular/certain/sentence.xls"
#>  [44] "dance/chair/shine/particular/certain/share.jpg"   
#>  [45] "dance/chair/shine/particular/certain/weight.jpg"  
#>  [46] "dance/chair/shine/particular/certain/short.xls"   
#>  [47] "dance/chair/shine/particular/certain/strong.doc"  
#>  [48] "dance/chair/shine/particular/certain/correct.doc" 
#>  [49] "dance/chair/shine/particular/certain/whose.xls"   
#>  [50] "dance/chair/shine/particular/window.pdf"          
#>  [51] "dance/chair/shine/particular/child.jpg"           
#>  [52] "dance/oxygen/drink/whose.doc"                     
#>  [53] "dance/oxygen/drink/earth.jpg"                     
#>  [54] "dance/oxygen/drink/after.pdf"                     
#>  [55] "dance/oxygen/drink/block/quart/brother.doc"       
#>  [56] "dance/oxygen/drink/block/quart/energy.pdf"        
#>  [57] "dance/oxygen/drink/block/quart/study.png"         
#>  [58] "dance/oxygen/drink/block/quart/heart.R"           
#>  [59] "dance/oxygen/drink/block/quart/drive.R"           
#>  [60] "dance/oxygen/drink/block/force.doc"               
#>  [61] "dance/oxygen/drink/block/friend.jpg"              
#>  [62] "dance/oxygen/drink/block/property/rather.pdf"     
#>  [63] "dance/oxygen/drink/block/property/collect.jpg"    
#>  [64] "dance/oxygen/drink/block/property/evening.xls"    
#>  [65] "dance/oxygen/drink/block/property/collect.png"    
#>  [66] "dance/oxygen/drink/block/property/exercise.pdf"   
#>  [67] "dance/oxygen/drink/block/supply.xls"              
#>  [68] "dance/oxygen/drink/block/capital/shall.xls"       
#>  [69] "dance/oxygen/drink/block/capital/shore.doc"       
#>  [70] "dance/oxygen/drink/block/capital/science.png"     
#>  [71] "dance/oxygen/drink/block/capital/record.R"        
#>  [72] "dance/oxygen/drink/block/capital/decide.xls"      
#>  [73] "dance/oxygen/drink/block/capital/continue.doc"    
#>  [74] "dance/oxygen/drink/block/capital/suffix.png"      
#>  [75] "dance/oxygen/drink/block/capital/search.jpg"      
#>  [76] "dance/oxygen/drink/block/capital/silent.xls"      
#>  [77] "dance/oxygen/drink/block/green/cover.jpg"         
#>  [78] "dance/oxygen/drink/block/green/perhaps.xls"       
#>  [79] "dance/oxygen/drink/shell.doc"                     
#>  [80] "dance/oxygen/drink/multiply.R"                    
#>  [81] "dance/oxygen/drink/imagine/seven.doc"             
#>  [82] "dance/oxygen/drink/imagine/chord.xls"             
#>  [83] "dance/oxygen/drink/imagine/summer.doc"            
#>  [84] "dance/oxygen/drink/imagine/bring/shell.xls"       
#>  [85] "dance/oxygen/drink/imagine/bring/gather.png"      
#>  [86] "dance/oxygen/drink/imagine/bring/could.xls"       
#>  [87] "dance/oxygen/drink/imagine/bring/string.jpg"      
#>  [88] "dance/oxygen/drink/imagine/bring/print.jpg"       
#>  [89] "dance/oxygen/drink/imagine/bring/where.jpg"       
#>  [90] "dance/oxygen/drink/imagine/bring/country.doc"     
#>  [91] "dance/oxygen/drink/imagine/particular.pdf"        
#>  [92] "dance/oxygen/could.png"                           
#>  [93] "dance/oxygen/danger.pdf"                          
#>  [94] "dance/oxygen/necessary.jpg"                       
#>  [95] "dance/single.R"                                   
#>  [96] "dance/their/protect/appear.pdf"                   
#>  [97] "dance/their/protect/guide/teach/finger.png"       
#>  [98] "dance/their/protect/guide/teach/death.doc"        
#>  [99] "dance/their/protect/guide/teach/question.R"       
#> [100] "dance/their/protect/guide/teach/laugh.R"          
#> [101] "dance/their/protect/guide/teach/between.R"        
#> [102] "dance/their/protect/guide/teach/shine.pdf"        
#> [103] "dance/their/protect/guide/teach/equal.jpg"        
#> [104] "dance/their/protect/guide/teach/interest.jpg"     
#> [105] "dance/their/protect/guide/teach/design.jpg"       
#> [106] "dance/their/protect/guide/teach/sight.pdf"        
#> [107] "dance/their/protect/guide/might.jpg"              
#> [108] "dance/their/protect/guide/listen/weather.jpg"     
#> [109] "dance/their/protect/guide/listen/father.pdf"      
#> [110] "dance/their/protect/guide/listen/million.pdf"     
#> [111] "dance/their/protect/guide/listen/engine.xls"      
#> [112] "dance/their/protect/guide/listen/every.R"         
#> [113] "dance/their/protect/guide/listen/thank.R"         
#> [114] "dance/their/protect/guide/listen/whose.R"         
#> [115] "dance/their/protect/guide/listen/summer.jpg"      
#> [116] "dance/their/protect/guide/listen/spring.png"      
#> [117] "dance/their/protect/guide/listen/occur.xls"       
#> [118] "dance/their/protect/guide/bright.R"               
#> [119] "dance/their/protect/guide/consider.xls"           
#> [120] "dance/their/protect/guide/determine.xls"          
#> [121] "dance/their/protect/guide/train.png"              
#> [122] "dance/their/protect/guide/stretch/clothe.jpg"     
#> [123] "dance/their/protect/guide/stretch/carry.jpg"      
#> [124] "dance/their/protect/guide/stretch/surprise.jpg"   
#> [125] "dance/their/protect/wrote.doc"                    
#> [126] "dance/their/protect/special.jpg"                  
#> [127] "dance/their/protect/class/woman.jpg"              
#> [128] "dance/their/protect/class/might/field.doc"        
#> [129] "dance/their/protect/class/might/solution.doc"     
#> [130] "dance/their/protect/class/might/apple.png"        
#> [131] "dance/their/protect/class/occur.pdf"              
#> [132] "dance/their/protect/class/meant.xls"              
#> [133] "dance/their/protect/class/quick.R"                
#> [134] "dance/their/protect/class/phrase.pdf"             
#> [135] "dance/their/protect/nature.doc"                   
#> [136] "dance/their/protect/interest.doc"                 
#> [137] "dance/their/earth.png"                            

# Frequency of path depths
table(lengths(kwb.file::split_paths(paths)))

#> Splitting paths ... ok. (0.00s) 

#> 
#>  2  3  4  5  6 
#>  1  4 13 23 96